+++ src/main/java/net/yacy/grid/io/messages/AbstractQueue.java
@@ -25,7 +25,7 @@ public abstract class AbstractQueue<A> implements Queue<A> {
 
     @Override
     public void clear() throws IOException {
-        int count = available();
+        long count = available();
         while (count-- > 0) {if  (receive(100) == null) break;}
     }
